# iCanteen-api接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "iCanteen-api接口",
    "description": "校园“智汇”食堂后端接口文档（返回统一使用 Result 包装）",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/window/list": {
      "get": {
        "summary": "免登录查询某食堂所有窗口",
        "deprecated": false,
        "description": "",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "canteenId",
            "in": "query",
            "description": "",
            "required": true,
            "example": 3,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=Window[])",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/window/{id}": {
      "get": {
        "summary": "免登录查询某窗口详细信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 3,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=Window)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除窗口（管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 9,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/window/{id}/dishes": {
      "get": {
        "summary": "免登录查询某窗口所有菜品",
        "deprecated": false,
        "description": "",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 4,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=Dish[])",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/window/report-wait-time": {
      "post": {
        "summary": "上报窗口排队时长",
        "deprecated": false,
        "description": "登录用户上报窗口排队时长。要求用户坐标到食堂坐标距离不超过300m；上报成功奖励20积分。",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "windowId",
            "in": "query",
            "description": "窗口ID",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "waitTime",
            "in": "query",
            "description": "排队时长（分钟）",
            "required": true,
            "example": 7,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "description": "用户经度",
            "required": true,
            "example": 120.149192,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "latitude",
            "in": "query",
            "description": "用户纬度",
            "required": true,
            "example": 30.316078,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "aa19515bcc324ccb811ac30548d0f520",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/window": {
      "post": {
        "summary": "增加窗口（管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowCreateDTO"
              },
              "example": {
                "canteenId": 3,
                "name": "麻辣烫",
                "description": "麻麻辣辣~",
                "sort": 0,
                "status": 1,
                "waitTime": 8
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result(data=id)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "修改窗口（管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "窗口"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowUpdateDTO"
              },
              "example": {
                "id": 9,
                "name": "你好西餐厅",
                "waitTime": 16
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMsg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "null"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
...
      "WindowCreateDTO": {
        "type": "object",
        "description": "管理员新增窗口参数（不需要id）",
        "properties": {
          "canteenId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "waitTime": {
            "type": "integer"
          },
          "waitTimeLevel": {
            "type": "integer"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "sort": {
            "type": "integer"
          }
        },
        "required": [
          "canteenId",
          "name"
        ]
      },
      "WindowUpdateDTO": {
        "type": "object",
        "description": "管理员修改窗口参数（必须携带id，其他字段按需修改）",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "canteenId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "waitTime": {
            "type": "integer"
          },
          "waitTimeLevel": {
            "type": "integer"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "sort": {
            "type": "integer"
          }
        },
        "required": [
          "id"
        ]
      },
...
    },
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}